moves cnl::_impl::encompasses into its own header
- moved out of cnl/num_traits.h
- it's mostly only needed by cnl/overflow.h
- moved tests from safe_integer.cpp into new encompasses.cpp
